Renamed repo card
